+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## 0.2.0 : 
 
 - **Feature**: Syslog client
-- **Change**: ARP refactor (`ArpManager` introduced)
+- **Feature**: Interface confugration
+- **Change**: ARP refactor (`ArpManager` with clock introduced)
 - **Change**: Format `PhysicalAddress` with colon
+- **Change**: `Device` entity renamed to `Interface`
 - **Update**: Updated PacketDotNet to 1.0.4 from 1.0.3
 
 ## 0.1.0 : 2020/02/03